handwriting-eval-api/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fastapi==0.104.1
fonttools==4.58.1
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-multipart==0.0.6
requests==2.32.4
scipy==1.15.3
six==1.17.0
uvicorn[standard]==0.24.0
